requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pynacl>=1.5.0
cryptography>=41.0.0
ecdsa>=0.18.0

